di: use booleans for axis
Since axis only has two possible values, represent it
with a boolean that can be used in if-statements. This
ensures that consumers only need to write an if-else
to cover all axes, instead of writing and if-elif-else
where the else case cannot happen.

e.g.:

if axis == Vertical {
} else if axis == Horzontal {
} else {
    // This can't happen
}

Becomes:

if axis == Vertical {
} else {
    // Must be horizontal
}

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>