di: add text progression types and method
This commit introduces a type that represents how a piece
of text is read relative to the origin as well as a
method to get the progression from any direction.

This information makes it easy to write algorithms that
operate on text of only one progression, but across any
axis.

Closes #14

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>