shaping: remove hacky direction heuristics
This commit drops heuristics for detecting the text direction of a
run of text in favor of explicitly requesting it. This makes the
code simpler, faster, and less prone to strange edge cases.

This commit also removes the need for mapRunesToClusterIndicies
to have access to the input text, as it only needs to know the
count of runes in the shaped text.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>