shaping: properly map runes at non-zero offsets
This commit updates mapRunesToClusterIndicies to correctly generate
mappings for ranges of glyphs that are not at the beginning of the
text. This is necessary because the harfbuzz cluster indices for each
glyph depend upon the range of runes they were shaped from. To replicate
the old behavior, I needed to normalize the cluster ids when comparing
them by subtracting the rune offset for the entire run of glyphs.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>