shaping: implement iterator wrap API
This commit creates a stand-alone function that can be used with a
state object to iteratively wrap a paragraph of text. This API allows
varying with width of the paragraph on a per-line basis (for use-cases
like inline images) and allows the caller to stop line-wrapping as soon
as they have all of the vertical text they need (for use-cases like
displaying the beginning of a paragraph with the rest off-screen).

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>