shaping: track candidate run separately
Previously, the algorithm would increment the persistent
iteration state's run index when it searched forward for
a better break candidate, but this is incorrect when part
of the current candidate's run is unused. If you end up
committing the current candidate, the next invocation of
WrapLine will skip the remainder of that run. This commit
updates the logic to avoid this mistake.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>