shaping: implement vertical text and drop error return
This commit alters the shaper to no longer return an error. The error return
was only used in a few cases, namely:

- trying to shape an unsupported text direction (vertical)
- shaping with invalid input parameters
- harfbuzz returning a glyph ID that doesn't exist within a font

The first of these can be resolved by implementing vertical text, which I
have done.

The second can be resolved by making reasonable choices about how to interpret
invalid input, and is also implemented.

The third cannot be fixed, but we can configure such glyphs to be invisible,
as such an error isn't user-recoverable anyway. This improves the robustness
of our shaper.

Now the shaper need not return an error.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>