shaping: implement custom truncator symbol
This commit enables the line wrapper to insert a custom symbol at the end
of a truncated line of text. This custom symbol is provided in the WrapConfig
as an already-shaped Output, allowing a great deal of flexibility in terms of
what glyph(s) are used. When shaping with a non-zero TruncateAfterLines, the
shaper will insert the Truncator Output at the end of the line unless the
line contains all remaining text to be shaped. The new truncated return value
from the wrapper provides the count of truncated runes, enabling applications
to easily determine which portion of the string is not visible.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>