Wrapping Refactors (#44)
* shaping: migrate mapping to new type

This change moves the operation of updating our current run's rune->glyph
mapping into a dedicated type. This removes the need for a closure within
the real line wrap algorithm, and is likely easier to understand also.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>

* shaping: migrate break selection to method

This commit moves the logic for choosing the next line breaking candidate into
a separate method of the wrapper. The process for choosing it is pretty independent
of the rest of the algorithm, and moving it makes the remaining code easier to read.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>

* shaping: remove incRun closure

This commit removes another closure from the line wrapping algorithm.
This closure barely improved the clarity of the code, and was an
impediment to other refactoring work.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>

* shaping: document nextBreakOption

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>

* shaping: extract slice resize operation

This commit moves the operation of resizing both the length and capacity
of a slice into a helper, making the code more readable and enabling us
to invoke the same logic elsewhere.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>

* shaping: move candidate commit to helper function

This commit migrates the logic for committing the best known line candidate
into a helper so that it can be used later.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>

---------

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>