shaping: allow forcing truncation
This commit enables application code to require the truncator to be inserted
when the text wraps to the maximum number of lines even when the text would
normally fit without truncation. Our API is paragraph-centric, and thus does
not know about paragraphs after itself. If trying to display truncatable text
with multiple paragraphs of contents, the previous API could not display a
truncator if the text wrapped a non-final paragraph to exactly fill the maximum
number of lines. This would result in the text not appearing to be truncated,
while it is missing entire lines of content.

The only reasonable solution I could see was to offer the option to always insert
the truncator when wrapping the final line. Application code can then track whether
there are future paragraphs (via readahead) and choose to request that the current
paragraph display as truncated if there are future paragraphs and this paragraph
fills all available lines.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>