shaping: [API] make wrapping consume iterator
This commit alters the line wrapping algorithm to consume an iterator that
provides shaped text, rather than the shaped text itself. This opens the
door for shaping text as the wrapping algorithm needs it to populate lines,
rather than doing all of the work in advance. Combined with the wrapper's
existing ability to truncate text, this means that the shape+wrap process
can terminate early once it has filled the requested number of lines.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>