shaping: [API] optimize wrapping allocations with caller-provided buffers
This commit creates the new shaping.WrapBuffer type and allows callers to
pass one into the wrapping process. This type can provide pre-allocated slices
for use within the line wrapping process, eliminating the need to allocate them
again. Providing these buffers greatly reduces allocations during line wrapping,
and generally provides a performance increase across the board when wrapping.

As the returned lines use the buffers as a backing store, it isn't safe to re-wrap
with the same buffer while still using the lines.

I've updated the benchmarks so that we compare our previous algorithm against the
performance of reusing buffers, as we're interested in optimal performance when
using the API correctly. However, I've also introduced a set of benchmarks suffixed
with -unbuffered that capture our performance when the caller does not provide a
line wrapping buffer.