shaping: implement grapheme boundary wrapping
This commit implements wrapping on grapheme cluster boundaries (UAX#29) in addition
to our previous word boundary (UAX#14) wrapping. The choice of whether to wrap on
grapheme clusters is left up to a new LineBreakPolicy type with three options:

- WhenNecessary: This option will break using grapheme clusters when no word can fit
on the line or when it is truncating a word (preserving as many grapheme clusters of
the word as possible). This is the default policy.
- Always: This option will break on grapheme clusters whenever possible, fitting as
many grapheme clusters as possible on every line.
- Never: This option disables the use of grapheme cluster breaking entirely, and
preserves the prior behavior.

As WhenNecessary is the zero value of LineBreakPolicy, consumers of our API will
experience a behavior change when updating to this version. Text that previously
overflowed its bounds because words were too long will now fit within them. I think
that this is going to be an improvement for 95% or more of use cases, so the change
is warranted.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>