shaping: fix breaker final grapheme iteration and tests
We need to allow the grapheme iterator to emit the very last rune's position
even though it is also always a word breaking candidate. Because the logic is
hierarchical, the word breaking process descends into the grapheme one when
policy dictates and there is no easy mechanism to say "nevermind, use what
the word breaker would have done". I don't think that can be implemented
efficiently.

When we're up against the very end of the text and the word breaker concludes
that the final run doesn't fit, it defers (most of the time) to the grapheme
breaker. The grapheme breaker would previously refuse to consider the end of
the text as a grapheme break because it knew the word breaker was already
considering that break. Now it handles this special case at the end of the
text.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>