shaping: discard break options from already-processed runs
Now that we can fall back to grapheme wrapping, an interesting edge case opens
up in which we have a very long UAX#14 segment composed of many runs and many
grapheme clusters. The word-oriented wrapping logic will fail to wrap it and
fall back to the grapheme-oriented logic, but the run iterator will be advanced
beyond some of the grapheme wrapping break options. Trying to process these
break options would crash the wrapper, so we now discard them.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>