shaping: move iterator state save/load up
This commit moves the management of the iterator cursor out of the break
candidate processing and into the loop that chooses which kind of break
candidate to use. This allows more intelligent choices about when to restore
the iterator state based on whether we are falling back to grapheme cluster
breaking or not.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>