fontscan: make cache dir configurable and change default
I've updated the font footprint cache to be stored by default in OS-specific
cache directories instead of application config ones. This does increase the
risk of the index file being deleted by some kind of regular cache clearing
process, but is a more appropriate location on the whole. Such deletion will
just cost some time on the next application startup.

I've also plumbed through a parameter allowing the caller to override the
cache directory when an application has need to.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>