fontscan: plumb through custom logging
This commit switches all logging callsites to use a caller-provided logger,
ensuring that the caller is able to control whether, where, and how font
scanning errors/warnings are logged. I've opted for a lack of logger to use
the default logger, but it would be equally reasonable for it to log to
io.Discard instead.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>