fontscan: add script enumeration method to runesets
This commit adds a mechanism to list the approximate set of covered scripts for
a rune set. I've opted to only test one rune per page in the set in the interest
of performance, so it's possible for this to miss certain covered scripts. It seems
better than nothing though.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>