fontscan: resolve faces by script coverage before giving up
This commit extends our fallback face resolution logic to check for fonts covering
the same script that the provided rune belongs. We first try only fonts with matching
aspects, and then we try all fonts with non-matching aspects. The non-matching case is
critical for making things like emoji work, as emoji rarely match the aspect of the text
containing them.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>