fontscan: initialize candidates after cache check
This commit moves the computation of the font candidates for the current query
to after we check for a cache hit. This saves an enormous quantity of redundant
work in real applications, as similar faces are frequently requested repeatedly,
and the information necessary to resolve them often lives entirely in the cache.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>