shaping: always modify run iterator and line candidate together
This commit fixes a class of bugs that arise when the current line candidate and
the run iterator become desynchronized. I hadn't previously consciously realized
the nature of the relationship, but these two states are fundamentally connected.
If the run iterator is ahead of the line candidate, we skip adding some runs to
candidate lines. If the run iterator is behind the line candidate, we duplicate
some runs within the line candidate. Whenever we save/restore one, we must also
save/restore the other. This commit captures that in some new operations on the
line wrapper in order to make doing so less error-prone.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>