shaping: (wrapping) make whitespace trim bidi-aware
Our previous implementation of trimming the whitespace off of the end of wrapped
lines assumed that the end was always the final logical run, and that the final
glyph within that run was visually last. Neither of these assumptions are true
in real-world bidi text. The final logical run may be embedded wtihin the center
of the visual line of text, and sometimes the final glyph is the first within a
run, rather than the last.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>