shaping: (wrapping) fix and stress test bidi run ordering
This commit adds a test for visual bidi run ordering and fixes the implementation
to pass it. The test itself is migrated from Gio's codebase, where it has been
verifying this functionality for years. It caught some problems with the new
implementation of this visual ordering algorithm, which are also fixed in this
commit.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>