shaping: (wrapping) fix space aware advance for bidi
When wrapping lines containing bidi runs, a run that progresses in the
opposite direction from the overall paragraph will have its trailing
whitespace in the middle of the line, not at the end. As such, it isn't
valid to ignore the advance of trailing whitespace for such a run. We
can only safely ignore trailing whitespace when the run is the same
progression as the paragraph.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>