shaping: treat truncator as proper run for bidi and rune counting
This commit fixes an oversight in our handling of text truncators. They were inserted
into the text *after* bidi run reordering had already occurred, which would place them
in the wrong visual position sometimes. Additionally, their rune data reflected the runes
of the truncator string instead of the runes of text the truncator represents, but it is
clearly more desirable for the truncator to stand in for all of the text it replaces.

This commit does change the semantics of a truncator run's Range field in the output text.
It will now describe the replaced range of text. This does greatly simplify consuming the
shaped text data, as you can see by how it simplified the tests. We don't have to factor
out the truncator or do extra math to make sure all runes are accounted for. Instead, the
sum of the rune counts of all runs will always equal the full rune count of the input
text.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>