shaping(wrapping): honor newline literals and other mandatory breaks
This commit updates our algorithm to respect literal newlines and other codepoints
indicating a mandatory line break. Unfortunately, doing so required removing some
fast-paths through the code because they worked by assuming that small quantities
of text could skip line wrapping entirely.

Now that the contents of the text can
force us to engage the line wrapping process in spite of whether the text would
fit without wrapping, we can't skip the line wrapping process as cheaply. I've
modified the most common codepath that processes single text runs so that it
scans the text for mandatory breaks and is able to take the fast path if none
are found, but this is more expensive than the old checks. Additionally, I've
removed a second fast path that was designed to skip the normal wrapping algorithm
if only one run remained to be wrapped. It was too expensive to properly scan
the remaining text for mandatory breaks in a way that didn't interfere with
the line wrapping algorithm.

Fixes #185

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>